Server layer: Change size limit to bytes to match heartbeat report value

+++ internal/server/master_server_types.go
@@ -13,5 +13,5 @@ type UploadResponse struct {
 
 type UpdateVolumeRequest struct {
 	VolumeID storage.VolumeID
-	Size     uint32
+	Size     uint64
 }
